Remove state_view mod
