+++ block-producer/src/lib.rs
@@ -2,5 +2,4 @@
 pub mod test_utils;
 
 pub mod batch_builder;
-pub mod state_view;
 pub mod txqueue;
