--- block-producer/src/state_view/mod.rs
@@ -1,13 +1,0 @@
-use async_trait::async_trait;
-use miden_objects::transaction::ProvenTransaction;
-use std::{fmt::Debug, sync::Arc};
-
-#[async_trait]
-pub trait TransactionVerifier: Send + Sync + 'static {
-    type VerifyTxError: Debug;
-
-    async fn verify_tx(
-        &self,
-        tx: Arc<ProvenTransaction>,
-    ) -> Result<(), Self::VerifyTxError>;
-}
