+++ block-producer/src/txqueue/mod.rs
@@ -1,16 +1,23 @@
-use std::{sync::Arc, time::Duration};
+use std::{fmt::Debug, sync::Arc, time::Duration};
 
 use async_trait::async_trait;
 use miden_objects::transaction::ProvenTransaction;
 use tokio::{sync::RwLock, time};
 
-use crate::{
-    batch_builder::{BatchBuilder, TransactionBatch},
-    state_view::TransactionVerifier,
-};
+use crate::batch_builder::{BatchBuilder, TransactionBatch};
 
 // TRANSACTION QUEUE
 // ================================================================================================
+
+#[async_trait]
+pub trait TransactionVerifier: Send + Sync + 'static {
+    type VerifyTxError: Debug;
+
+    async fn verify_tx(
+        &self,
+        tx: Arc<ProvenTransaction>,
+    ) -> Result<(), Self::VerifyTxError>;
+}
 
 #[async_trait]
 pub trait TransactionQueue: Send + Sync + 'static {
